Finishing touches in gallery, moving onto album
